Made a simple dynamic page for each song from the songs.json file
