+++ app/music/Navigation.jsx
@@ -1,3 +1,5 @@
+"use client"
+
 import styles from './Navigation.module.css';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
