+++ app/music/[id]/page.module.css
@@ -1,0 +1,53 @@
+.container {
+ 	backdrop-filter: blur(100px);
+}
+
+.background {
+	position: fixed;
+	top: 0;
+	left: 0;
+	width: 100%;
+	height: 100%;
+	background-color: rgba(0, 0, 0, 0.5);
+	z-index: -1;
+	filter: blur(100px);
+	background-repeat: no-repeat;
+	background-size: cover;
+}
+
+.main {
+	display: grid;
+	grid-template-columns: auto 1fr 1fr;
+	grid-template-rows: 1fr 1fr;
+	gap: 15px;
+	width: 90%;
+	height: 80%;
+	margin: auto;
+	margin-top: 80px;
+}
+
+.main :global(.SongCover) {
+	border: 1px solid var(--border-color);
+}
+
+.title {
+ margin-top: 10px;
+ text-wrap: nowrap;
+}
+
+@media (max-width: 425px) { 
+	.main {
+		grid-template-columns: 1fr;
+		grid-template-rows: 1fr 1fr;
+	}
+
+	.main :global(.SongCover) {
+		margin: auto;
+	}
+
+	.title {
+		font-size: 1.5em;
+		text-align: center;
+		margin-top: 20px;
+	}
+}
